@@ -6,20 +6,20 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.10.5 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  provider: ^4.3.3
-  gap: ^1.2.0
+  provider: ^5.0.0
+  gap: ^2.0.0-nullsafety.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  effective_dart: ^1.3.0
+  effective_dart: ^1.3.1
 
 flutter:
   uses-material-design: false
