@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 
 class Screen1 extends StatelessWidget {
-  const Screen1({Key key}) : super(key: key);
+  const Screen1({Key? key}) : super(key: key);
 
   static const routeName = '/screen_1';
 
