@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 
 class Screen2 extends StatelessWidget {
-  const Screen2({Key key}) : super(key: key);
+  const Screen2({Key? key}) : super(key: key);
 
   static const routeName = '/screen_2';
 
